requirements.txt
fastapi==0.77.1 
python-dotenv==0.20.0
requests==2.28.1
uvicorn==0.18.2